# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: siwx-oidc
  version: 0.2.0
  summary: A CAIP-122 → OpenID Connect bridge, and the OIDC provider for a Matrix homeserver.
  description: |
    siwx-oidc turns a wallet signature or a passkey into standard OIDC tokens.
    Any relying party can use it; the primary deployment target is Matrix
    Synapse, where it is the auth service in Synapse's
    `matrix_authentication_service` integration (the Matrix OAuth 2.0 API,
    MSC3861).

    # Identity model

    Every user carries three identifiers with three different owners, and
    conflating any two of them is a bug class rather than a style question:

    | Tier | Value | Owner | Mutable |
    |---|---|---|---|
    | Alias | a generated `Firstname Surname` pseudonym | the user | yes |
    | MXID | `@{base36(sha256(did)[..10])}:{server}` | derived | no — Synapse has no rename API |
    | DID | `did:key:…` / `did:pkh:…` plus the provider's signature | the provider | no |

    # Authentication schemes in this document

    Four different credentials appear below, and they are not interchangeable:

    * **bearerToken** — an access token this provider issued, presented by the
      end user's client.
    * **masSharedSecret** — the shared secret of Synapse's
      `matrix_authentication_service` integration. Server-to-server only;
      the homeserver holds it. It must never traverse the public internet.
    * **registrationAccessToken** — returned by dynamic client registration and
...

...

...

...
  license:
...
servers:
...
paths:
...
  /oauth2/introspect:
    post:
      tags: [Homeserver]
      operationId: introspect
      summary: Token introspection (RFC 7662) — MSC3861
      description: |
        Called by Synapse to validate the access token on authenticated
        requests (Synapse caches a successful answer, `active: false` included,
        for two minutes). The response's `scope` is what Synapse's
        `is_server_admin()` reads, so it is also how an admin-scoped token gets
        its privilege.

        `device_id` is rendered as JSON `null`, never `""`: Synapse 1.159
        rejects a present-but-empty device id with a 500.
      security: [{ masSharedSecret: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties: { token: { type: string } }
      responses:
        "200":
          description: |
            Token metadata, or `{"active": false}`. An UNKNOWN token is
            `{"active": false}` with HTTP 200 — never a 4xx.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Introspection" }
        "404": { description: Introspection is unavailable — no `mas_shared_secret` is configured. }
  /oauth2/admin_token:
    post:
      tags: [Homeserver]
      operationId: adminToken
      summary: Mint a short-lived Synapse-admin-scoped token
      description: |
        Synapse 1.157 deleted the `admin_token` shim, so `/_synapse/admin/*` is
        reachable only with a token whose introspected scope carries
        `urn:synapse:admin:*`. This mints one.

        TTL is clamped **in code** to 30..=900s so a large
        `SIWXOIDC_ADMIN_TOKEN_TTL_SECS` cannot promote it into a standing admin
        key. Minting idempotently provisions `SIWXOIDC_ADMIN_TOKEN_LOCALPART`,
        which creates a real Matrix account.
      security: [{ masSharedSecret: [] }]
      responses:
        "200":
          description: The minted token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminToken" }
        "401": { description: Wrong or missing shared secret. }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: An access token issued by this provider (`mat_…` in delegated-auth mode).
    masSharedSecret:
      type: http
      scheme: bearer
      description: |
        The `matrix_authentication_service` shared secret, compared for exact
        string equality.
        Server-to-server only — the homeserver holds it. Never expose these
        routes to the public internet.
    registrationAccessToken:
      type: http
      scheme: bearer
      description: The registration access token returned by `/register`, scoped to that client.
    sessionCookie:
      type: apiKey
      in: cookie
      name: siwx
      description: |
        A browser ceremony cookie. `siwx` carries the CAIP-122 proof,
        `siwx_user` is an opaque passkey-picker scope hint, and `acct_session`
        scopes the account page. These are interactive flows, not an API.
...
  schemas:
...
    Introspection:
      type: object
      properties:
        active: { type: boolean }
        username: { type: string, description: The Matrix localpart. }
        device_id: { type: [string, "null"], description: "`null`, never `\"\"` — an empty string is a 500 at Synapse." }
        scope: { type: string }
        sub: { type: string, description: The DID. }
    AdminToken:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string }
        expires_in: { type: integer, description: Clamped in code to 30..=900. }
        scope: { type: string, description: "Must carry BOTH `urn:matrix:client:api:*` and `urn:synapse:admin:*`." }
        user_id: { type: string }
...